March23

+++ UsersSubscriptions/Areas/Teacher/Controllers/TeacherController.cs
@@ -12,7 +12,12 @@ namespace UsersSubscriptions.Areas.Teacher.Controllers
     public class TeacherController : Controller
     {
         private ITeacherRepository repository;
-        public TeacherController(ITeacherRepository repo) => repository = repo;
+        private IUserRepository userRepository;
+        public TeacherController(ITeacherRepository repo, IUserRepository userRepo)
+        {
+            repository = repo;
+            userRepository = userRepo;
+        }
     
         public async Task<IActionResult> Index()
         {
@@ -32,11 +37,13 @@ namespace UsersSubscriptions.Areas.Teacher.Controllers
             Course course = await repository.GetCoursInfoAsync(Id);
             return View(course);
         }
+
         public async Task<IActionResult> ConfirmSubscription(string Id)
         {
             Subscription subscription = await repository.GetSubscriptionAsync(Id);
             return View(subscription);
         }
+
         [HttpPost]
         public async Task<IActionResult> ConfirmSubscription(Subscription subsc)
         {
@@ -65,8 +72,83 @@ namespace UsersSubscriptions.Areas.Teacher.Controllers
         [HttpPost]
         public async Task<IActionResult> RemoveSubscription(Subscription subscription)
         {
-            await repository.RemoveSubscriptionAsync(subscription.Id);
+            if (await repository.GetSubscriptionAsync(subscription.Id) != null)
+            {
+                await repository.RemoveSubscriptionAsync(subscription.Id);
+            }
             return RedirectToAction(nameof(CourseInfo),new { Id=subscription.CourseId});
+        }
+
+       // [HttpPost]
+        public async Task<IActionResult> StudentInfo(string studentQR)
+        {
+            AppUser student = await repository.GetUserAsync(studentQR);
+            if (student == null)
+            {
+                return RedirectToAction(nameof(ScanQrCode));
+            }
+            IEnumerable<Course> teacherCourses = await repository.GetTeacherCoursesAsync(await repository.GetCurrentUserAsync(HttpContext));
+            teacherCourses = teacherCourses.Where(cour => cour.IsActive == true);
+            StudentInfoViewModel model = new StudentInfoViewModel
+            {
+                Student = student,
+                Courses = teacherCourses
+            };
+            return View(model);
+        }
+
+        public async Task<IActionResult> CourseStudentInfo(string studentId, string courseId)
+        {
+            IEnumerable<Subscription> studentSubscriptions = repository.GetStudentSubscriptionsOfCourse(studentId, courseId);
+            AppUser student = await repository.GetUserAsync(studentId);
+            Course course = await repository.GetCoursInfoAsync(courseId);
+            CourceStudentViewModel model = new CourceStudentViewModel
+            {
+                Course = course,
+                Student = student,
+                Subscriptions = studentSubscriptions
+            };
+            return View(model);
+        }
+
+        public async Task<IActionResult> AddSubscription(string studentId, string courseId)
+        {
+            SubscriptionTeatcherViewModel model = new SubscriptionTeatcherViewModel
+            {
+                Student = await repository.GetUserAsync(studentId),
+                Course = await repository.GetCoursInfoAsync(courseId)
+            };
+            return View(model);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> AddSubscription(SubscriptionTeatcherViewModel model)
+        {
+            AppUser teacher = await repository.GetCurrentUserAsync(HttpContext);
+            Subscription subscription = new Subscription {
+                AppUserId = model.Student.Id,
+                ConfirmedDatetime = DateTime.Now,
+                CreatedDatetime = DateTime.Now,
+                ConfirmedById = teacher.Id,
+                CourseId = model.Course.Id,
+                DayStart=model.Subscription.DayStart,
+        };
+            await userRepository.CreateSubscription(subscription);
+            return RedirectToAction(nameof(StudentInfo), new { studentQR = model.Student.Id});
+        }
+
+         public async Task<IActionResult> RemoveStudentSubscription(string Id)
+        {
+            Subscription subscription = await repository.GetSubscriptionAsync(Id);
+            return View(subscription);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> RemoveStudentSubscription(Subscription model)
+        {
+            Subscription subscription = await repository.GetSubscriptionAsync(model.Id);
+            await repository.RemoveSubscriptionAsync(model.Id);
+            return RedirectToAction(nameof(StudentInfo), new { studentQR = subscription.AppUserId });
         }
     }
 }
